HBASE-23245 : MutableHistogram constructor changes and provide HistogramImpl maxExpected as long (#787)

+++ hbase-metrics/src/main/java/org/apache/hadoop/hbase/metrics/impl/HistogramImpl.java
@@ -36,7 +36,7 @@ public class HistogramImpl implements Histogram {
   private final CounterImpl counter;
 
   public HistogramImpl() {
-    this(Integer.MAX_VALUE << 2);
+    this((long) Integer.MAX_VALUE << 2);
   }
 
   public HistogramImpl(long maxExpected) {
